Invite Only
Move register form to logged in only, password must be set via reset/forgot. add dockerfile for debug email server
