+++ app/auth/forms.py
@@ -14,24 +14,6 @@ class LoginForm(FlaskForm):
     submit = SubmitField(_l('Sign in'))
 
 
-class RegistrationFrom(FlaskForm):
-    username = StringField(_l('Username'), validators=[DataRequired()])
-    email = StringField(_l('Email'), validators=[Email(), DataRequired()])
-    password = PasswordField(_l('Password'), validators=[DataRequired()])
-    password2 = PasswordField(_l('Repeat Password'), validators=[DataRequired(), EqualTo('password')])
-    submit = SubmitField(_l('Register'))
-
-    def validate_username(self, username):
-        user = db.session.scalar(sa.select(User).where(User.username == username.data))
-        if user is not None:
-            raise ValidationError(_l('Please use a different username'))
-
-    def validate_email(self, email):
-        user = db.session.scalar(sa.select(User).where(User.email == email.data))
-        if user is not None:
-            raise ValidationError(_l('Please use a different email'))
-
-
 class ResetPasswordRequestForm(FlaskForm):
     email = StringField(_l('Email'), validators=[DataRequired(), Email()])
     submit = SubmitField(_l('Request Password Reset'))
