+++ app/templates/auth/login.html
@@ -8,8 +8,4 @@
         {{ _('Forgot Your Password?') }}
         <a href="{{ url_for('auth.reset_password_request') }}">{{ _('Click to Reset It') }}</a>
     </p>
-    <p>
-        {{ _('New User?') }}
-        <a href="{{ url_for('auth.register') }}">{{ _('Click to Register!') }}</a>
-    </p>
 {% endblock %}
