+++ app/templates/base.html
@@ -60,6 +60,9 @@
                     <a class="nav-link" aria-current="page" href="{{ url_for('main.user', username=current_user.username) }}">{{ _('Profile') }}</a>
                 </li>
                 <li class="nav-item">
+                    <a class="nav-link" aria-current="page" href="{{ url_for('main.register') }}">{{ _('Invite') }}</a>
+                </li>
+                <li class="nav-item">
                     <a class="nav-link" aria-current="page" href="{{ url_for('auth.logout') }}">{{ _('Logout') }}</a>
                 </li>
                 {% endif %}
