+++ app/templates/register.html
@@ -2,6 +2,6 @@
 {% import "bootstrap_wtf.html" as wtf %}
 
 {% block content %}
-    <h1>{{ _('Register') }}</h1>
+    <h1>{{ _('Invite') }}</h1>
     {{ wtf.quick_form(form) }}
 {% endblock %}
